updated settings

+++ src/services/apiCabins.js
@@ -11,21 +11,32 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
   );
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  // https://hzoyaxzijespjlithyjc.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
+  //1. Create/Edit Cabin
+  let query = supabase.from("cabins");
 
-  //1. Create Cabin
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  //A)CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  //B) EDIT
+  if (id) {
+    query = query
+      .update({ ...newCabin, image: imagePath })
+      .eq("id", id)
+      .select();
+  }
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
@@ -34,11 +45,13 @@ export async function createCabin(newCabin) {
 
   //2. Upload image
 
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
-  //3. Delete the cabin IF there was an error uploadinf the corresponding image
+  //3. Delete the cabin IF there was an error uploading the corresponding image
 
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
